SR-Unet-tf2/SR-UNet/requirements.txt
scipy==1.4.1
numpy==1.18.4
matplotlib==3.2.1
opencv_python==4.2.0.34
tensorflow_gpu==2.4.0
tqdm==4.46.1
Pillow==8.2.0
h5py==2.10.0
